Ensure generation of export output file parent folder(s)
